+++ src/MetaheuristicsPlatform/Algorithms/PSO/BareBones/BareBonesPsoReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.PSO.BareBones;
+
+public static class BareBonesPsoReferences
+{
+    public static ScientificReference Kennedy2003 { get; } =
+        new(
+            "Kennedy",
+            2003,
+            "Bare bones particle swarms",
+            "Proceedings of the 2003 IEEE Swarm Intelligence Symposium, 80-87",
+            "10.1109/SIS.2003.1202251");
+}